requirements.txt
discord.py
requests
newspaper3k